+++ Python/Product/Analysis/LanguageServer/DelayedAction.cs
@@ -1,0 +1,34 @@
+﻿// Python Tools for Visual Studio
+// Copyright(c) Microsoft Corporation
+// All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the License); you may not use
+// this file except in compliance with the License. You may obtain a copy of the
+// License at http://www.apache.org/licenses/LICENSE-2.0
+//
+// THIS CODE IS PROVIDED ON AN  *AS IS* BASIS, WITHOUT WARRANTIES OR CONDITIONS
+// OF ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING WITHOUT LIMITATION ANY
+// IMPLIED WARRANTIES OR CONDITIONS OF TITLE, FITNESS FOR A PARTICULAR PURPOSE,
+// MERCHANTABLITY OR NON-INFRINGEMENT.
+//
+// See the Apache Version 2.0 License for specific language governing
+// permissions and limitations under the License.
+
+using System;
+using System.Threading;
+
+namespace Microsoft.PythonTools.Analysis.LanguageServer {
+    sealed class DelayedAction {
+        private readonly Timer _timer;
+        private bool _canceled;
+        public DelayedAction(Action action, int timeout) {
+            _timer = new Timer((s) => {
+                if (!_canceled) {
+                    action();
+                }
+                _timer?.Dispose();
+            }, null, timeout, Timeout.Infinite);
+        }
+        public void Cancel() => _canceled = true;
+    }
+}
