+++ Python/Product/Analysis/LanguageServer/WorkspaceSymbolsHandler.cs
@@ -1,0 +1,109 @@
+﻿// Python Tools for Visual Studio
+// Copyright(c) Microsoft Corporation
+// All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the License); you may not use
+// this file except in compliance with the License. You may obtain a copy of the
+// License at http://www.apache.org/licenses/LICENSE-2.0
+//
+// THIS CODE IS PROVIDED ON AN  *AS IS* BASIS, WITHOUT WARRANTIES OR CONDITIONS
+// OF ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING WITHOUT LIMITATION ANY
+// IMPLIED WARRANTIES OR CONDITIONS OF TITLE, FITNESS FOR A PARTICULAR PURPOSE,
+// MERCHANTABLITY OR NON-INFRINGEMENT.
+//
+// See the Apache Version 2.0 License for specific language governing
+// permissions and limitations under the License.
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using Microsoft.PythonTools.Analysis.Infrastructure;
+using Microsoft.PythonTools.Interpreter;
+
+namespace Microsoft.PythonTools.Analysis.LanguageServer {
+    internal sealed class WorkspaceSymbolsHandler {
+        private readonly ProjectFiles _projectFiles;
+        public WorkspaceSymbolsHandler(ProjectFiles projectFiles) {
+            _projectFiles = projectFiles;
+        }
+
+        public SymbolInformation[] GetWorkspaceSymbols(WorkspaceSymbolParams @params) {
+            var members = Enumerable.Empty<MemberResult>();
+            var opts = GetMemberOptions.ExcludeBuiltins | GetMemberOptions.DeclaredOnly;
+
+            foreach (var entry in _projectFiles.All) {
+                members = members.Concat(
+                    GetModuleVariables(entry as IPythonProjectEntry, opts, @params.query)
+                );
+            }
+
+            members = members.GroupBy(mr => mr.Name).Select(g => g.First());
+            return members.Select(m => ToSymbolInformation(m)).ToArray();
+        }
+
+        private static IEnumerable<MemberResult> GetModuleVariables(
+            IPythonProjectEntry entry,
+            GetMemberOptions opts,
+            string prefix
+        ) {
+            var analysis = entry?.Analysis;
+            if (analysis == null) {
+                yield break;
+            }
+
+            foreach (var m in analysis.GetAllAvailableMembers(SourceLocation.None, opts)) {
+                if (m.Values.Any(v => v.DeclaringModule == entry)) {
+                    if (string.IsNullOrEmpty(prefix) || m.Name.StartsWithOrdinal(prefix, ignoreCase: true)) {
+                        yield return m;
+                    }
+                }
+            }
+        }
+
+        private SymbolInformation ToSymbolInformation(MemberResult m) {
+            var res = new SymbolInformation {
+                name = m.Name,
+                kind = ToSymbolKind(m.MemberType),
+                _kind = m.MemberType.ToString().ToLowerInvariant()
+            };
+
+            var loc = m.Locations.FirstOrDefault(l => !string.IsNullOrEmpty(l.FilePath));
+            if (loc != null) {
+                res.location = new Location {
+                    uri = new Uri(PathUtils.NormalizePath(loc.FilePath), UriKind.RelativeOrAbsolute),
+                    range = new SourceSpan(
+                        new SourceLocation(loc.StartLine, loc.StartColumn),
+                        new SourceLocation(loc.EndLine ?? loc.StartLine, loc.EndColumn ?? loc.StartColumn)
+                    )
+                };
+            }
+
+            return res;
+        }
+
+        private SymbolKind ToSymbolKind(PythonMemberType memberType) {
+            switch (memberType) {
+                case PythonMemberType.Unknown: return SymbolKind.None;
+                case PythonMemberType.Class: return SymbolKind.Class;
+                case PythonMemberType.Instance: return SymbolKind.Object;
+                case PythonMemberType.Delegate: return SymbolKind.Function;
+                case PythonMemberType.DelegateInstance: return SymbolKind.Function;
+                case PythonMemberType.Enum: return SymbolKind.Enum;
+                case PythonMemberType.EnumInstance: return SymbolKind.EnumMember;
+                case PythonMemberType.Function: return SymbolKind.Function;
+                case PythonMemberType.Method: return SymbolKind.Method;
+                case PythonMemberType.Module: return SymbolKind.Module;
+                case PythonMemberType.Namespace: return SymbolKind.Namespace;
+                case PythonMemberType.Constant: return SymbolKind.Constant;
+                case PythonMemberType.Event: return SymbolKind.Event;
+                case PythonMemberType.Field: return SymbolKind.Field;
+                case PythonMemberType.Property: return SymbolKind.Property;
+                case PythonMemberType.Multiple: return SymbolKind.Object;
+                case PythonMemberType.Keyword: return SymbolKind.None;
+                case PythonMemberType.CodeSnippet: return SymbolKind.None;
+                case PythonMemberType.NamedArgument: return SymbolKind.None;
+                default: return SymbolKind.None;
+            }
+        }
+    }
+}
